requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
PyYAML==6.0.2
Jinja2==3.1.6
python-multipart==0.0.20
packaging==24.2
Markdown==3.7
